apidump: Rewrite codegen to use VulkanObject
Major rewrite of the api_dump_generator.py script to use VulkanObject
instead of its own OutputGenerator subclass. This rewrite attemps to
change as little as possible in the generated code, to aid in validating
the rewrite. But doing so introduced a few small errors as the default
sorting was disabled. Subsequent commits will enable sorting and clean
up the introduced mistakes.